refactored demo servlets
