+++ backend/src/main/java/eu/chessdata/backend/api/DemoServlet02.java
@@ -1,0 +1,18 @@
+package eu.chessdata.backend.api;
+
+import java.io.IOException;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+/**
+ * Created by Bogdan Oloeriu on 03/07/2016.
+ */
+public class DemoServlet02 extends HttpServlet{
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        super.doPost(req, resp);
+    }
+}
