+++ backend/src/main/webapp/WEB-INF/appengine-web.xml
@@ -1,10 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <appengine-web-app xmlns="http://appengine.google.com/ns/1.0">
-    <application>myApplicationId</application>
-    <version>1</version>
+    <application>chess-data</application>
+    <version>2</version>
     <threadsafe>true</threadsafe>
-
     <system-properties>
-        <property name="java.util.logging.config.file" value="WEB-INF/logging.properties"/>
+        <property name="java.util.logging.config.file" value="WEB-INF/logging.properties" />
     </system-properties>
 </appengine-web-app>
