+++ backend/src/main/webapp/WEB-INF/web.xml
@@ -2,7 +2,7 @@
 <web-app xmlns="http://java.sun.com/xml/ns/javaee" version="2.5">
     <servlet>
         <servlet-name>MyServlet</servlet-name>
-        <servlet-class>eu.chessdata.backend.MyServlet</servlet-class>
+        <servlet-class>eu.chessdata.backend.api.DemoServlet01</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>MyServlet</servlet-name>
